Restructured groups to add authenticated view
