+++ src/app/[sdSlug]/groups/details/[groupId]/components/AuthenticatedView.tsx
@@ -1,0 +1,6 @@
+"use client";
+
+
+export function AuthenticatedView() {
+  return <>Logged In</>
+}
